Add files via upload
An incomplete version of the initial survey for the Flag project. This test is missing the strings and flag inputs and is missing the settings page. I am having some difficulty with the settings config, which I need to solve before I can implement the randomly generated strings, and the flag images.